Update on "Move Method::lowered_graph to a separate pass out of the Method class."
There is no reason to have it in a general API of Module/Method - it's
just another graph pass. It was there because some time ago modules were
not first class and all graphs were lowered. After that changed, this
API was added for easier transition, but now we don't need it anymore.

Differential Revision: [D17986724](https://our.internmc.facebook.com/intern/diff/D17986724)

[ghstack-poisoned]
