+++ tools/build_variables.py
@@ -122,6 +122,7 @@ libtorch_sources = [
     "torch/csrc/jit/passes/liveness.cpp",
     "torch/csrc/jit/passes/loop_unrolling.cpp",
     "torch/csrc/jit/passes/lower_grad_of.cpp",
+    "torch/csrc/jit/passes/lower_graph.cpp",
     "torch/csrc/jit/passes/lower_tuples.cpp",
     "torch/csrc/jit/passes/peephole.cpp",
     "torch/csrc/jit/passes/python_print.cpp",
